Get some debugging information on various potential problems when making partials

Get some debugging information on various potential problems when making partials

knit_child_debug(...)

Arguments

...
passed to partial()

Examples

knit_child_debug()

#> dev.new(): using pdf(file="Rplots4.pdf")

#> dev.new(): using pdf(file="Rplots5.pdf")

#> dev.new(): using pdf(file="Rplots6.pdf")

#> dev.new(): using pdf(file="Rplots7.pdf")

#> dev.new(): using pdf(file="Rplots8.pdf")

#> dev.new(): using pdf(file="Rplots9.pdf")

#> dev.new(): using pdf(file="Rplots10.pdf")

#> No viewer found, probably checking

#> 
#> <details>
#> 
#> <summary>Debug</summary>
#> 
#> Working directory
#> ## [1] "/Users/rubenarslan/research/rmarkdownpartials/docs/reference"
#> 
#> 
#> Knitr Output.dir
#> ## [1] "/Users/rubenarslan/research/rmarkdownpartials/docs/reference"
#> 
#> 
#> Child mode
#> ## [1] TRUE
#> 
#> 
#> 
#> Viewer null
#> ## [1] TRUE
#> 
#> 
#> in temp dir
#> ## [1] FALSE
#> 
#> 
#> <details>
#> 
#> <summary>opts_knit</summary>
#> ## $progress
#> ## [1] FALSE
#> ## 
#> ## $verbose
#> ## [1] FALSE
#> ## 
#> ## $width
#> ## [1] 75
#> ## 
#> ## $eval.after
#> ## [1] "fig.cap"
#> ## 
#> ## $base.dir
#> ## NULL
#> ## 
#> ## $base.url
#> ## NULL
#> ## 
#> ## $root.dir
#> ## NULL
#> ## 
#> ## $child.path
#> ## [1] ""
#> ## 
#> ## $upload.fun
#> ## function (x) 
#> ## x
#> ## <bytecode: 0x7fe04b8084f8>
#> ## <environment: namespace:base>
#> ## 
#> ## $global.device
#> ## [1] FALSE
#> ## 
#> ## $global.par
#> ## [1] FALSE
#> ## 
#> ## $concordance
#> ## [1] FALSE
#> ## 
#> ## $documentation
#> ## [1] 1
#> ## 
#> ## $self.contained
#> ## [1] TRUE
#> ## 
#> ## $unnamed.chunk.label
#> ## [1] "unnamed-chunk"
#> ## 
#> ## $highr.opts
#> ## NULL
#> ## 
#> ## $out.format
#> ## [1] "markdown"
#> ## 
#> ## $child
#> ## [1] TRUE
#> ## 
#> ## $parent
#> ## [1] FALSE
#> ## 
#> ## $tangle
#> ## [1] FALSE
#> ## 
#> ## $aliases
#> ## NULL
#> ## 
#> ## $header
#> ## highlight      tikz    framed 
#> ##        ""        ""        "" 
#> ## 
#> ## $global.pars
#> ## NULL
#> ## 
#> ## $output.dir
#> ## [1] "/Users/rubenarslan/research/rmarkdownpartials/docs/reference"
#> 
#> 
#> </details>
#> 
#> 
#> <details>
#> 
#> <summary>opts_chunk</summary>
#> ## $eval
#> ## [1] TRUE
#> ## 
#> ## $echo
#> ## [1] FALSE
#> ## 
#> ## $results
#> ## [1] "markup"
#> ## 
#> ## $tidy
#> ## [1] FALSE
#> ## 
#> ## $tidy.opts
#> ## NULL
#> ## 
#> ## $collapse
#> ## [1] FALSE
#> ## 
#> ## $prompt
#> ## [1] FALSE
#> ## 
#> ## $comment
#> ## [1] "##"
#> ## 
#> ## $highlight
#> ## [1] TRUE
#> ## 
#> ## $strip.white
#> ## [1] TRUE
#> ## 
#> ## $size
#> ## [1] "normalsize"
#> ## 
#> ## $background
#> ## [1] "#F7F7F7"
#> ## 
#> ## $cache
#> ## [1] FALSE
#> ## 
#> ## $cache.path
#> ## [1] "cache/a538d47a5567e879b0b59ce21b4c0b00_"
#> ## 
#> ## $cache.vars
#> ## NULL
#> ## 
#> ## $cache.lazy
#> ## [1] TRUE
#> ## 
#> ## $dependson
#> ## NULL
#> ## 
#> ## $autodep
#> ## [1] FALSE
#> ## 
#> ## $cache.rebuild
#> ## [1] FALSE
#> ## 
#> ## $fig.keep
#> ## [1] "high"
#> ## 
#> ## $fig.show
#> ## [1] "asis"
#> ## 
#> ## $fig.align
#> ## [1] "default"
#> ## 
#> ## $fig.path
#> ## [1] "figure/a538d47a5567e879b0b59ce21b4c0b00_"
#> ## 
#> ## $dev
#> ## NULL
#> ## 
#> ## $dev.args
#> ## NULL
#> ## 
#> ## $dpi
#> ## [1] 72
#> ## 
#> ## $fig.ext
#> ## NULL
#> ## 
#> ## $fig.width
#> ## [1] 7
#> ## 
#> ## $fig.height
#> ## [1] 7
#> ## 
#> ## $fig.env
#> ## [1] "figure"
#> ## 
#> ## $fig.cap
#> ## NULL
#> ## 
#> ## $fig.scap
#> ## NULL
#> ## 
#> ## $fig.lp
#> ## [1] "fig:"
#> ## 
#> ## $fig.subcap
#> ## NULL
#> ## 
#> ## $fig.pos
#> ## [1] ""
#> ## 
#> ## $out.width
#> ## NULL
#> ## 
#> ## $out.height
#> ## NULL
#> ## 
#> ## $out.extra
#> ## NULL
#> ## 
#> ## $fig.retina
#> ## [1] 1
#> ## 
#> ## $external
#> ## [1] TRUE
#> ## 
#> ## $sanitize
#> ## [1] FALSE
#> ## 
#> ## $interval
#> ## [1] 1
#> ## 
#> ## $aniopts
#> ## [1] "controls,loop"
#> ## 
#> ## $warning
#> ## [1] TRUE
#> ## 
#> ## $error
#> ## [1] TRUE
#> ## 
#> ## $message
#> ## [1] TRUE
#> ## 
#> ## $render
#> ## NULL
#> ## 
#> ## $ref.label
#> ## NULL
#> ## 
#> ## $child
#> ## NULL
#> ## 
#> ## $engine
#> ## [1] "R"
#> ## 
#> ## $split
#> ## [1] FALSE
#> ## 
#> ## $include
#> ## [1] TRUE
#> ## 
#> ## $purl
#> ## [1] TRUE
#> 
#> 
#> </details>
#> 
#> </details>