Generate a small plot that will be enlarged in a modal when clicked

Generate a small plot that will be enlarged in a modal when clicked

enlarge_plot(
  plot,
  large_plot = plot,
  plot_name = NULL,
  width_small = 2,
  height_small = 2,
  width_large = 7,
  height_large = 7,
  ...
)

Arguments

plot
a plot

large_plot
a larger version of the same plot. defaults to the first plot if left empty, but this only works for ggplot2 and similar, not base plots

plot_name
optional: specify a meaningful plot name (needs to be unique in the document)

width_small
width for the small plot

height_small
height for the small plot

width_large
width for the large plot

height_large
height for the large plot

...
passed to partial()

Examples

# will generate figures in a temporary directory
if (requireNamespace("ggplot2")) {
dist <- ggplot2::qplot(stats::rbeta(200, 3, 4))
enlarge_plot(dist,
large_plot = dist + ggplot2::theme_classic(base_size = 18))
} else {
graphics::hist(stats::rbeta(200, 3, 4))
dist <- grDevices::recordPlot()
enlarge_plot(dist)
}

#> Loading required namespace: ggplot2

#> dev.new(): using pdf(file="Rplots1.pdf")

#> dev.new(): using pdf(file="Rplots2.pdf")

#> No viewer found, probably checking

#> 
#> 
#> 
#> 
#> <!-- Button trigger modal -->
#> <a href="#bfe5bb17abfcd06e3509aa76d5316a5d" data-toggle="modal" data-target="#bfe5bb17abfcd06e3509aa76d5316a5d">
#> ## `stat_bin()` using `bins = 30`. Pick better value with `binwidth`.
#> figure/bfe5bb17abfcd06e3509aa76d5316a5d_rmdpartialchunk-1-1.pdf
#> </a>
#> 
#> <!-- Modal -->
#> <div class="modal fade" id="bfe5bb17abfcd06e3509aa76d5316a5d" tabindex="-1" role="dialog" aria-labelledby="bfe5bb17abfcd06e3509aa76d5316a5dLabel" aria-hidden="true">
#> <div class="modal-dialog" role="document">
#> <div class="modal-content">
#> <div class="modal-body">
#> 
#> ## `stat_bin()` using `bins = 30`. Pick better value with `binwidth`.
#> figure/bfe5bb17abfcd06e3509aa76d5316a5d_rmdpartialchunk-2-1.pdf
#> 
#> </div>
#> <div class="modal-footer">
#> <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
#> </div>
#> </div>
#> </div>
#> </div>
#> 
#>